+++ packages/expo-observe/src/module.ts
@@ -3,6 +3,7 @@ import AppMetrics, { setErrorHandlerEnabled } from 'expo-app-metrics';
 
 import { initRouterIntegration } from './integrations/expo-router/init';
 import { isRouterInstalled } from './integrations/expo-router/router';
+import { initReanimatedIntegration } from './integrations/react-native-reanimated/init';
 import { initReactNavigationIntegration } from './integrations/react-navigation/init';
 import { isReactNavigationInstalled } from './integrations/react-navigation/reactNavigation';
 import { reportCaughtError } from './reportCaughtError';
@@ -79,6 +80,9 @@ const Observe: ObserveModule = new Proxy(native, {
           initRouterIntegration(config.integrations?.['expo-router']);
         } else if (shouldInitReactNavigationIntegration) {
           initReactNavigationIntegration(config.integrations?.['react-navigation']);
+        }
+        if (config.integrations?.['react-native-reanimated']) {
+          initReanimatedIntegration();
         }
         return target.configure(config);
       };